Separate components by concern
- 각 컴포넌트들을 별도의 파일로 분리했습니다.
- '관심사'에 따른 분리라는 개념이 아직도 어려운 것 같습니다.
