+++ src/App.jsx
@@ -1,0 +1,40 @@
+import React, { useState } from 'react';
+import Title from './Title';
+import Default from './Default';
+import Form from './Form';
+import List from './List';
+
+function App() {
+  const [task, setTodo] = useState({ value: '' });
+  function handleChange(event) {
+    setTodo({ value: event.target.value });
+  }
+
+  const [todoList, setTodoList] = useState([]);
+  function handleSubmit(event) {
+    event.preventDefault();
+    if (task.value === '') return;
+
+    setTodoList([...todoList, { value: task.value }]);
+    setTodo({ value: '' });
+    document.getElementById('todoinput').value = '';
+  }
+
+  function handleClickDone(index) {
+    return setTodoList(
+      [...todoList.slice(0, index), ...todoList.slice(index + 1, todoList.length)],
+    );
+  }
+
+  return (
+    <div>
+      <Title />
+      <Form onChange={handleChange} onSubmit={handleSubmit} />
+      {todoList.length === 0
+        ? (<Default />)
+        : (<List values={[...todoList]} onClick={handleClickDone} />)}
+    </div>
+  );
+}
+
+export default App;
