+++ src/ButtonDone.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function ButtonDone({ onClick }) {
+  return <button type="button" onClick={onClick}>완료</button>;
+}
+
+export default ButtonDone;
