+++ src/Default.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Default() {
+  return <p>할 일이 없어요!</p>;
+}
+
+export default Default;
