+++ src/Submit.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Submit() {
+  return <input type="submit" value="추가" />;
+}
+
+export default Submit;
