+++ src/Task.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import ButtonDone from './ButtonDone';
+
+function Task({ dataKey, value, onClick }) {
+  return (
+    <li>
+      {value}
+      <ButtonDone onClick={() => onClick(dataKey)} />
+    </li>
+  );
+}
+
+export default Task;
