+++ src/Title.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Title() {
+  return <h1>To-do</h1>;
+}
+
+export default Title;
